+++ pkg/autoscaler/autoscaler/autoscaler_manager.go
@@ -119,26 +119,16 @@ func (am *autoScalerManager) syncAutoScaling(tc *v1alpha1.TidbCluster, tac *v1al
 	// Request PD for auto-scaling plans
 	plans, err := controller.GetPDClient(am.pdControl, tc).GetAutoscalingPlans(*strategy)
 	if err != nil {
-		klog.Errorf("cannot get auto-scaling plans %v", err)
+		klog.Errorf("cannot get auto-scaling plans err:%v", err)
 		return err
 	}
 
-	// Distinguish plans
-	plansMap := make(map[string][]pdapi.Plan)
-	for _, plan := range plans {
-		plansMap[plan.Component] = append(plansMap[plan.Component], plan)
+	// Apply auto-scaling plans
+	if err := am.syncPlans(tc, tac, plans); err != nil {
+		klog.Errorf("cannot apply autoscaling plans err:%v", err)
+		return err
 	}
 
-	oldTikvReplicas := tc.Spec.TiKV.Replicas
-	if err := am.syncTiKV(tc, tac); err != nil {
-		tc.Spec.TiKV.Replicas = oldTikvReplicas
-		klog.Errorf("tac[%s/%s] tikv sync failed, continue to sync next, err:%v", tac.Namespace, tac.Name, err)
-	}
-	oldTidbReplicas := tc.Spec.TiDB.Replicas
-	if err := am.syncTiDB(tc, tac); err != nil {
-		tc.Spec.TiDB.Replicas = oldTidbReplicas
-		klog.Errorf("tac[%s/%s] tidb sync failed, continue to sync next, err:%v", tac.Namespace, tac.Name, err)
-	}
 	klog.Infof("tc[%s/%s]'s tac[%s/%s] synced", tc.Namespace, tc.Name, tac.Namespace, tac.Name)
 	return nil
 }
@@ -164,8 +154,7 @@ func (am *autoScalerManager) syncTidbClusterReplicas(tac *v1alpha1.TidbClusterAu
 	return nil
 }
 
-func (am *autoScalerManager) updateAutoScaling(oldTc *v1alpha1.TidbCluster,
-	tac *v1alpha1.TidbClusterAutoScaler) error {
+func (am *autoScalerManager) updateAutoScaling(oldTc *v1alpha1.TidbCluster, tac *v1alpha1.TidbClusterAutoScaler) error {
 	if tac.Annotations == nil {
 		tac.Annotations = map[string]string{}
 	}
